Evals/LLM Arena/requirements.txt
click
pyyaml
faiss-cpu
streamlit
PyPDF2
numpy

# LLM API libraries
google-generativeai
openai
anthropic

# Additional dependencies for LLM libraries
google-api-core
google-auth
google-auth-httplib2
google-api-python-client
google-ai-generativelanguage
googleapis-common-protos
grpcio
grpcio-status
proto-plus
protobuf
rsa
pyasn1
pyasn1-modules
pydantic
pydantic-core
annotated-types
typing-inspection
tqdm
uritemplate
httplib2
pyparsing
anyio
distro
httpx
jiter
sniffio
httpcore
h11